feat: 회원가입 및 정보 수정 로직 수정
