+++ backend/src/main/java/com/metamong/mt/domain/member/dto/request/ConsumerSignUpRequestDto.java
@@ -6,6 +6,7 @@ import com.metamong.mt.domain.member.dto.request.validation.EnumValidator;
 import com.metamong.mt.domain.member.model.Member;
 import com.metamong.mt.domain.member.model.constant.Gender;
 import com.metamong.mt.domain.member.model.constant.Role;
+import com.metamong.mt.global.constant.BooleanAlt;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
@@ -70,6 +71,7 @@ public class ConsumerSignUpRequestDto {
                 .memPostalCode(this.memPostalCode)
                 .memAddress(this.memAddress)
                 .memDetailAddress(this.memDetailAddress)
+                .isDel(BooleanAlt.N)
                 .role(Role.ROLE_CONS)
                 .build();
     }
