+++ backend/src/main/java/com/metamong/mt/domain/member/mailformatter/IdNotificationMailMessageFormatter.java
@@ -1,4 +1,4 @@
-package com.metamong.mt.domain.member.service.mailformatter;
+package com.metamong.mt.domain.member.mailformatter;
 
 import org.springframework.stereotype.Component;
 
@@ -6,17 +6,17 @@ import com.metamong.mt.global.mail.MailMessageFormatter;
 import com.metamong.mt.global.mail.MailType;
 
 @Component
-public class MailValidationMailMessageFormatter extends MailMessageFormatter {
+public class IdNotificationMailMessageFormatter extends MailMessageFormatter {
 
     @Override
     protected String getMessageFormat() {
         return """
-                인증 번호는 "{0}"입니다.
+                회원님의 아이디는 {0} 입니다.
                 """;
     }
 
     @Override
     public MailType getSupportedMailType() {
-        return MailType.MAIL_VALIDATION;
+        return MailType.ID_NOTIFICATION;
     }
 }
