+++ backend/src/main/java/com/metamong/mt/domain/member/mailformatter/PasswordResetLinkMailMessageFormatter.java
@@ -1,4 +1,4 @@
-package com.metamong.mt.domain.member.service.mailformatter;
+package com.metamong.mt.domain.member.mailformatter;
 
 import org.springframework.stereotype.Component;
